criacao main

+++ Entrega2/Valores.java
@@ -1,13 +1,13 @@
 package OOP;
 public class Valores {
-	private int valorFracao,
+	private float valorFracao,
 				valorCheio,
 				valorDiaria,
 				valorNoturno,
 				valorMensal,
 				valorEvento;
 	private Valores() {}
-	public Valores(int v1, int v2, int v3, int v4, int v5, int v6) {
+	public Valores(float v1, float v2, float v3, float v4, float v5, float v6) {
 		this.valorFracao = v1;
 		this.valorCheio = v2;
 		this.valorDiaria = v3;
@@ -19,42 +19,42 @@ public class Valores {
 	public int getValorFracao() {
 		return valorFracao;
 	}
-	public void setValorFracao(int v) {
+	public void setValorFracao(float v) {
 		this.valorFracao = v;
 	}
 	
 	public int getValorCheio() {
 		return valorCheio;
 	}
-	public void setValorCheio(int v) {
+	public void setValorCheio(float v) {
 		this.valorCheio = v;
 	}
 	
 	public int getValorDiaria() {
 		return valorDiaria;
 	}
-	public void setValorDiaria(int v) {
+	public void setValorDiaria(float v) {
 		this.valorDiaria = v;
 	}
 	
 	public int getValorNoturno() { 
 		return valorNoturno;
 	}
-	public void setValorNoturno(int v) {
+	public void setValorNoturno(float v) {
 		this.valorNoturno = v;
 	}
 	
 	public int getValorMensal() {
 		return valorMensal;
 	}
-	public void setValorMensal(int v) {
+	public void setValorMensal(float v) {
 		this.valorMensal = v;
 	}
 	
 	public int getValorEvento() {
 		return valorEvento;
 	}
-	public void setValorEvento(int v) {
+	public void setValorEvento(float v) {
 		this.valorEvento = v;
 	}
